@@ -5,30 +5,29 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.*;
 import org.model.transactions.Transaction;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.core.io.DefaultResourceLoader;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.ResourceLoader;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.context.ActiveProfiles;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
 import org.transactions.persistence.model.TransactionES;
 import org.transactions.persistence.repositories.TransactionAggregateRepository;
+import org.transactions.transactionssyncprocess.service.AppStartService;
 import org.transactions.transactionssyncprocess.utils.TransactionsElasticsearchContainer;
 import org.transactions.transactionssyncprocess.utils.TransactionsMongoDbContainer;
 
+import java.io.IOException;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @Testcontainers
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureMockMvc
+@ActiveProfiles("test")
 public class IntegrationTest {
 
     @Container
@@ -52,7 +51,7 @@ public class IntegrationTest {
     private TransactionAggregateRepository aggregateRepository;
 
     @Autowired
-    private MockMvc mvc;
+    AppStartService startService;
 
     @BeforeAll
     static void setUp() {
@@ -85,19 +84,23 @@ public class IntegrationTest {
         // Prepare Data - Insert Data into mongodb database
 
         // Get sample JSON file
-        Resource resource = resourceLoader.getResource("classpath:data/sample_file.json");
-        // Convert it into Transanction object
-        List<Transaction> transactions = mapper.readValue(resource.getFile(), mapper.getTypeFactory().constructCollectionType(List.class, Transaction.class));
-        // Persist them into database
-        transactions.forEach(item -> mongoTemplate.save(item, "transaction"));
+        loadTransactionsInBDD();
 
         // Run sync service
-        mvc.perform(post("/sync")).andExpect(status().isNoContent());
+        startService.start();
 
         // Assertions - records must exist in ES database
         Iterable<TransactionES> result = aggregateRepository.findAll();
         Assertions.assertTrue(result.iterator().hasNext());
 
+    }
+
+    private void loadTransactionsInBDD() throws IOException {
+        Resource resource = resourceLoader.getResource("classpath:data/sample_file.json");
+        // Convert it into Transanction object
+        List<Transaction> transactions = mapper.readValue(resource.getFile(), mapper.getTypeFactory().constructCollectionType(List.class, Transaction.class));
+        // Persist them into database
+        transactions.forEach(item -> mongoTemplate.save(item, "transaction"));
     }
 
     @AfterAll
