@@ -4,7 +4,7 @@ import org.testcontainers.containers.MongoDBContainer;
 
 public class TransactionsMongoDbContainer extends MongoDBContainer {
 
-    private static final String MONGODB_IMAGE = "mongo:4.0.10";
+    private static final String MONGODB_IMAGE = "mongo:6.0.24";
 
     public TransactionsMongoDbContainer() {
         super(MONGODB_IMAGE);
